+++ src/sheen.rs
@@ -1,4 +1,4 @@
-use crate::{Attribute, EffectType, DescriptionFormat};
+use crate::{Attribute, AttributeValue, EffectType, DescriptionFormat};
 use strum::{Display, EnumString, EnumIter, EnumCount};
 use num_enum::{TryFromPrimitive, IntoPrimitive};
 use serde_repr::{Serialize_repr, Deserialize_repr};
@@ -50,4 +50,14 @@ impl Attribute for Sheen {
     const EFFECT_TYPE: EffectType = EffectType::Positive;
     const HIDDEN: bool = false;
     const STORED_AS_INTEGER: bool = false;
+    
+    /// Gets the attribute value.
+    fn attribute_value(&self) -> Option<AttributeValue> {
+        None
+    }
+    
+    /// Gets the attribute float value.
+    fn attribute_float_value(&self) -> Option<f64> {
+        Some((*self as u32) as f64)
+    }
 }